Updated schema management to use SchemaDefinition organization
