+++ py-modules/core/macrostrat/core/__init__.py
@@ -4,3 +4,4 @@ app = Macrostrat()
 
 # This has to happen after the macrostrat config
 from .database import get_database  # noqa
+from .schema_definition import SchemaDefinition  # noqa
